track: Update road (2)

+++ configure.sh
@@ -87,7 +87,15 @@ if [ -d "$source" ]; then
 
     path="$PWD/track/$1/dist/upscale"
 
-    cp "$source"/*.jpeg "$path"
+    if exists "$source"/*.jpeg; then
+
+        cp "$source"/*.jpeg "$path"
+    fi
+
+    if exists "$source"/*.mp4; then
+
+        cp "$source"/*.mp4 "$path"
+    fi
 fi
 
 source="$assets/track/$1/dist/upscale/base"
